* 6.3:
  always pass microseconds to usleep as integers
  Bump Symfony version to 6.3.11
  Update VERSION for 6.3.10
  Update CHANGELOG for 6.3.10
  Bump Symfony version to 5.4.34
  Update VERSION for 5.4.33
  Update CONTRIBUTORS for 5.4.33
  Update CHANGELOG for 5.4.33

+++ src/Symfony/Component/HttpClient/Response/TransportResponseTrait.php
@@ -296,7 +296,7 @@ trait TransportResponseTrait
             }
 
             if (-1 === self::select($multi, min($timeoutMin, $timeoutMax - $elapsedTimeout))) {
-                usleep(min(500, 1E6 * $timeoutMin));
+                usleep((int) min(500, 1E6 * $timeoutMin));
             }
 
             $elapsedTimeout = hrtime(true) / 1E9 - $lastActivity;
